# docs/swagger.yaml
definitions:
...
  models.VmsTrnReceivedKeyDriver:
    properties:
      trn_request_uid:
        example: 3045a994-ba0b-431d-acf2-98768a9c5fc9
        type: string
    type: object
  models.VmsTrnReceivedKeyOutSider:
    properties:
      outsider_name:
        example: John Doe
        type: string
      received_key_mobile_contact_number:
        example: "0812345678"
        type: string
      received_key_remark:
        example: OutSider received the key
        type: string
      trn_request_uid:
        example: 3045a994-ba0b-431d-acf2-98768a9c5fc9
        type: string
    type: object
  models.VmsTrnReceivedKeyPEA:
    properties:
      received_key_emp_id:
        example: "990001"
        type: string
      received_key_internal_contact_number:
        example: "5551234"
        type: string
      received_key_mobile_contact_number:
        example: "0812345678"
        type: string
      received_key_remark:
        example: Employee received the key
        type: string
      trn_request_uid:
        example: 3045a994-ba0b-431d-acf2-98768a9c5fc9
        type: string
    type: object
...
  models.VmsTrnRequestCanceled:
    properties:
      canceled_request_reason:
        example: Test Cancel
        type: string
      trn_request_uid:
        example: 8bd09808-61fa-42fd-8a03-bf961b5678cd
        type: string
    type: object
...
  models.VmsTrnRequestUpdateRecieivedKey:
    properties:
      received_key_end_datetime:
        example: "2025-02-16T17:00:00Z"
        type: string
      received_key_place:
        example: Main Office
        type: string
      received_key_start_datetime:
        example: "2025-02-16T08:00:00Z"
        type: string
      trn_request_uid:
        example: 8bd09808-61fa-42fd-8a03-bf961b5678cd
        type: string
    type: object
...
  models.VmsTrnRequestUpdateRecieivedKeyDetail:
    properties:
      received_key_datetime:
        example: "2025-02-16T08:00:00Z"
        type: string
      ref_vehicle_key_type_code:
        example: 1
        type: integer
      trn_request_uid:
        example: 8bd09808-61fa-42fd-8a03-bf961b5678cd
        type: string
    type: object
...
info:
  contact: {}
  title: VMS_PLUS
  version: "1.0"
paths:
...
  /api/received-key-admin/request/{trn_request_uid}:
    get:
      consumes:
      - application/json
      description: This endpoint fetches details of a specific booking request using
        its unique identifier (TrnRequestUID).
      parameters:
      - description: TrnRequestUID (trn_request_uid)
        in: path
        name: trn_request_uid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Retrieve a specific booking request
      tags:
      - Received-key-admin
  /api/received-key-admin/search-requests:
    get:
      consumes:
      - application/json
      description: Search for requests using a keyword and get the summary of counts
        grouped by request status code
      parameters:
      - description: Search keyword (matches request_no, vehicle_license_plate, vehicle_user_emp_name,
          or work_place)
        in: query
        name: search
        type: string
      - description: Filter by multiple request status codes (comma-separated, e.g.,
          'A,B,C')
        in: query
        name: ref_request_status_code
        type: string
      - description: Filter by start datetime (YYYY-MM-DD format)
        in: query
        name: startdate
        type: string
      - description: Filter by end datetime (YYYY-MM-DD format)
        in: query
        name: enddate
        type: string
      - description: Filter by vehicle owner department SAP
        in: query
        name: vehicle_owner_dept_sap
        type: string
      - description: Filter by received key start datetime (YYYY-MM-DD format)
        in: query
        name: received_key_start_datetime
        type: string
      - description: Filter by received key end datetime (YYYY-MM-DD format)
        in: query
        name: received_key_end_datetime
        type: string
      - description: Order by request_no, start_datetime, ref_request_status_code
        in: query
        name: order_by
        type: string
      - description: 'Order direction: asc or desc'
        in: query
        name: order_dir
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of records per page (default: 10)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Search booking requests and get summary counts by request status code
      tags:
      - Received-key-admin
  /api/received-key-admin/update-canceled:
    put:
      consumes:
      - application/json
      description: This endpoint allows users to update the cancel status of an item.
      parameters:
      - description: VmsTrnRequestCanceled data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnRequestCanceled'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update cancel status for an item
      tags:
      - Received-key-admin
  /api/received-key-admin/update-key-pickup-driver:
    put:
      consumes:
      - application/json
      description: This endpoint allows to update key pickup driver.
      parameters:
      - description: VmsTrnReceivedKeyDriver data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnReceivedKeyDriver'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update key pickup driver for a booking request
      tags:
      - Received-key-admin
  /api/received-key-admin/update-key-pickup-outsider:
    put:
      consumes:
      - application/json
      description: This endpoint allows to update key pickup outsource.
      parameters:
      - description: VmsTrnReceivedKeyOutSider data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnReceivedKeyOutSider'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update key pickup outsource for a booking request
      tags:
      - Received-key-admin
  /api/received-key-admin/update-key-pickup-pea:
    put:
      consumes:
      - application/json
      description: This endpoint allows to update key pickup emp user.
      parameters:
      - description: VmsTrnReceivedKeyPEA data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnReceivedKeyPEA'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update key pickup emp user for a booking request
      tags:
      - Received-key-admin
  /api/received-key-admin/update-recieived-key:
    put:
      consumes:
      - application/json
      description: This endpoint allows to update key pickup driver.
      parameters:
      - description: VmsTrnRequestUpdateRecieivedKey data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnRequestUpdateRecieivedKey'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update key pickup driver for a booking request
      tags:
      - Received-key-admin
  /api/received-key-admin/update-recieived-key-detail:
    put:
      consumes:
      - application/json
      description: This endpoint allows confirming the update of the key pickup driver
        for a specific booking request.
      parameters:
      - description: VmsTrnRequestUpdateRecieivedKeyDetail data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnRequestUpdateRecieivedKeyDetail'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Confirm the update of key pickup driver for a booking request
      tags:
      - Received-key-admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-ApiKey
    type: apiKey
  AuthorizationAuth:
    description: Bearer [your Authorization]
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"